# Replaces `faer` `concat!` macro with block Hamiltonian function and makes minor refactors to the Rust code