This PR fixes #77 by replacing the `faer` `concat!` macro, which is not recommended for external use, with a more stable own function implementation. This doesn't significantly change the function runtime:
```
old concat macro:
+------+------------------------+-----------------+--------------------+------------------------+------------------+
| n_q  | heisenberg_ferromagnet | antiferro_chain | kagome_ferromagnet | kagome_antiferromagnet | kagome_supercell |
+------+------------------------+-----------------+--------------------+------------------------+------------------+
| 100  |        0.007045        |     0.006178    |      0.013309      |        0.016799        |     0.194486     |
| 1000 |        0.028488        |     0.029554    |      0.098945      |        0.126499        |     1.914994     |
| 5000 |        0.134682        |     0.141327    |      0.507639      |        0.659204        |    12.774269     |
+------+------------------------+-----------------+--------------------+------------------------+------------------+
new block hamiltonian function:
+------+------------------------+-----------------+--------------------+------------------------+------------------+
| n_q  | heisenberg_ferromagnet | antiferro_chain | kagome_ferromagnet | kagome_antiferromagnet | kagome_supercell |
+------+------------------------+-----------------+--------------------+------------------------+------------------+
| 100  |        0.005785        |     0.005534    |      0.013091      |        0.014965        |     0.189698     |
| 1000 |        0.027250        |     0.030260    |      0.094214      |        0.121221        |     1.879567     |
| 5000 |        0.130085        |     0.144967    |      0.484230      |        0.613518        |    12.258631     |
+------+------------------------+-----------------+--------------------+------------------------+------------------+
```

this PR also makes some minor refactors to the Rust code:
- changes the internal functions to not have a preceding `_` (Rust denotes public functions with the `pub fn` syntax, so this isn't necessary and just makes the code harder to read for no reason)
- adds the #[inline] macro to internal functions (which can create some very minor performance improvements with basically no downside)
- avoids needing to re-iterate over vectors so much by removing an unnecessary iteration in the Python bindings over q-vectors, and combining the 3 calls to `get_rotation_component` into one `get_rotation_components` function
    - the `get_rotation_components` function also now avoids copying the 3rd component of the rotation matrices (eta) when it's only ever read
- changes some Vec's into slices in the single-Q spinwave function, which apparently helps the compiler optimise it...